Re-arrange helper files - Rename helpers.ts to generic.ts - Put .ts and test.ts file in same folder - Undo removed typing in roundTo
